refactor: move SandboxedIframe into application package
