--- packages/iframe/package.json
@@ -1,29 +1,0 @@
-{
-  "name": "@bos-web-engine/iframe",
-  "version": "1.0.0",
-  "description": "",
-  "main": "lib/index.js",
-  "scripts": {
-    "build": "tsc -p ./tsconfig.json",
-    "dev": "tsc -w -p ./tsconfig.json",
-    "lint": "eslint ./src/**/*.ts*",
-    "lint:fix": "eslint . --fix"
-  },
-  "dependencies": {
-    "@bos-web-engine/container": "workspace:*"
-  },
-  "devDependencies": {
-    "@bos-web-engine/common": "workspace:*",
-    "@types/node": "^17.0.12",
-    "@types/react": "^18.0.17",
-    "@types/react-dom": "^18.0.6",
-    "eslint": "^7.32.0",
-    "eslint-config-custom": "workspace:*",
-    "react": "^18.2.0",
-    "tsconfig": "workspace:*",
-    "typescript": "^4.5.2"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC"
-}
