--- packages/iframe/src/index.ts
@@ -1,2 +1,0 @@
-export { SandboxedIframe } from './SandboxedIframe';
-export { getAppDomId, getIframeId } from './utils';
