--- packages/iframe/src/utils.ts
@@ -1,7 +1,0 @@
-export function getAppDomId(id: string) {
-  return `dom-${id}`;
-}
-
-export function getIframeId(id: string) {
-  return `iframe-${id}`;
-}
